Project/electricity-provider/requirements.txt
Django==4.2.7
environs==9.5.0
psycopg2
djangorestframework
django-cors-headers
python-dotenv
requests
django-bootstrap-v5
